refactor: create login route, update home page
